Merge from main to devops/documentation

+++ Vendor-Link-Point/Controllers/JateksController.cs
@@ -1,15 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Vendor_Link_Point.Data;
 using Vendor_Link_Point.Models;
 
 namespace Vendor_Link_Point.Controllers
 {
+    [Authorize(Roles = "Kereskedo")]
     public class JateksController : Controller
     {
         private readonly VendorLinkPointContext _context;
@@ -22,23 +24,28 @@ namespace Vendor_Link_Point.Controllers
         // GET: Jateks
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Jatek.ToListAsync());
+            var myKereskedoId = User.FindFirst("KereskedoId")?.Value;
+
+            // 2. SZŰRÉS: Csak a bejelentkezett kereskedő könyveit kérjük le
+            var myJateks = await _context.Jatek
+                .Where(j => j.KereskedoId == myKereskedoId)
+                .ToListAsync();
+
+            return View(myJateks);
         }
 
         // GET: Jateks/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
+            var myKereskedoId = User.FindFirst("KereskedoId")?.Value;
+
+            // Csak akkor mutatjuk meg, ha az övé
             var jatek = await _context.Jatek
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (jatek == null)
-            {
-                return NotFound();
-            }
+                .FirstOrDefaultAsync(m => m.Id == id && m.KereskedoId == myKereskedoId);
+
+            if (jatek == null) return NotFound();
 
             return View(jatek);
         }
@@ -56,8 +63,17 @@ namespace Vendor_Link_Point.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Korhatar,Tipus,Id,Nev,Gyarto,Ar,Raktarkeszlet,Kategoria,Leiras,KepUrl,Elerheto")] Jatek jatek)
         {
+            ModelState.Remove("KereskedoId");
+            ModelState.Remove("Kategoria");
+            ModelState.Remove("Elerheto");
+
             if (ModelState.IsValid)
             {
+                // 3. HOZZÁRENDELÉS: Mentés előtt rögzítjük, hogy ki a tulajdonos
+                jatek.KereskedoId = User.FindFirst("KereskedoId")?.Value;
+                jatek.Kategoria = "Játékok"; // Biztosíték
+                jatek.Elerheto = jatek.Raktarkeszlet > 0;
+
                 _context.Add(jatek);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -68,16 +84,18 @@ namespace Vendor_Link_Point.Controllers
         // GET: Jateks/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var jatek = await _context.Jatek.FindAsync(id);
-            if (jatek == null)
+            if (jatek == null) return NotFound();
+
+            // 4. JOGOSULTSÁG ELLENŐRZÉS: Tényleg ő a tulajdonos?
+            var myKereskedoId = User.FindFirst("KereskedoId")?.Value;
+            if (jatek.KereskedoId != myKereskedoId)
             {
-                return NotFound();
+                return Unauthorized(); // Ha másét akarja szerkeszteni, elutasítjuk!
             }
+
             return View(jatek);
         }
 
@@ -88,28 +106,28 @@ namespace Vendor_Link_Point.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Korhatar,Tipus,Id,Nev,Gyarto,Ar,Raktarkeszlet,Kategoria,Leiras,KepUrl,Elerheto")] Jatek jatek)
         {
-            if (id != jatek.Id)
-            {
-                return NotFound();
-            }
+            if (id != jatek.Id) return NotFound();
+
+            var myKereskedoId = User.FindFirst("KereskedoId")?.Value;
+
+            ModelState.Remove("KereskedoId");
+            ModelState.Remove("Kategoria");
 
             if (ModelState.IsValid)
             {
                 try
                 {
+                    // Visszapótoljuk a rejtett azonosítót, nehogy elvesszen a frissítéskor
+                    jatek.KereskedoId = myKereskedoId;
+                    jatek.Kategoria = "Játékok";
+
                     _context.Update(jatek);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!JatekExists(jatek.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!JatekExists(jatek.Id)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -119,17 +137,14 @@ namespace Vendor_Link_Point.Controllers
         // GET: Jateks/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
+
+            var myKereskedoId = User.FindFirst("KereskedoId")?.Value;
 
             var jatek = await _context.Jatek
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (jatek == null)
-            {
-                return NotFound();
-            }
+                .FirstOrDefaultAsync(m => m.Id == id && m.KereskedoId == myKereskedoId);
+
+            if (jatek == null) return NotFound(); // Ha másé, úgy teszünk, mintha nem is létezne
 
             return View(jatek);
         }
@@ -139,13 +154,16 @@ namespace Vendor_Link_Point.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var myKereskedoId = User.FindFirst("KereskedoId")?.Value;
             var jatek = await _context.Jatek.FindAsync(id);
-            if (jatek != null)
+
+            // Törlés előtt is meggyőződünk róla, hogy az övé
+            if (jatek != null && jatek.KereskedoId == myKereskedoId)
             {
                 _context.Jatek.Remove(jatek);
+                await _context.SaveChangesAsync();
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
